Se solucionaron conflictos en la rama Modularizacion_base
