+++ api/index.js
@@ -1,4 +1,4 @@
-export { getUserById } from './usuarios/crearUsuario/getUserById.js';
+export { getUserById } from './usuarios/getUserById.js';
 export { getTareasByUserId } from './tareas/getTareasByUserId.js';
 export { createTarea } from './tareas/createTarea.js';
 export { deleteTarea } from './tareas/deleteTarea.js';
