--- use-case/usuarios/ciudades/getCiudades.js
@@ -1,5 +1,0 @@
-export const ciudades = async () => {
-  const solicitud = await fetch('http://localhost:3001/ciudades');
-  const datos = await solicitud.json();
-  return datos;
-}
