--- use-case/usuarios/crearUsuario/getUserById.js
@@ -1,8 +1,0 @@
-export const getUserById = async (idUsuario) => {
-    const respuesta = await fetch(`https://jsonplaceholder.typicode.com/users/${idUsuario}`);
-    if (respuesta.ok) {
-        return await respuesta.json();
-    } else {
-        throw new Error("Usuario no encontrado");
-    }
-}
