--- use-case/usuarios/generos/getGeneros.js
@@ -1,5 +1,0 @@
-export const generos = async () => {
-  const solicitud = await fetch('http://localhost:3001/generos');
-  const datos = await solicitud.json();
-  return datos;
-}
